+++ domain/src/main/java/com/wanted/domain/repository/CompanyRepository.kt
@@ -4,5 +4,5 @@ import com.wanted.domain.model.CompanyInfoModel
 import com.wanted.domain.result.DomainResult
 
 interface CompanyRepository {
-        suspend fun getCompanyDetail(companyId: Long): DomainResult<CompanyInfoModel>
+        suspend fun getCompanyDetail(companyId: Int): DomainResult<CompanyInfoModel>
 }
